+++ udata/commands/tests/test_fixtures.py
@@ -7,7 +7,9 @@ import werkzeug.test
 from pytest_mock import MockerFixture
 from werkzeug.wrappers.response import Response
 
+import udata.commands.fixtures
 from udata import models
+from udata.core.dataservices.factories import DataserviceFactory
 from udata.core.dataset.factories import (
     CommunityResourceFactory,
     DatasetFactory,
@@ -42,6 +44,7 @@ class FixturesTest:
             user=user,
             discussion=[MessageDiscussionFactory(**{}, posted_by=user)],
         )
+        DataserviceFactory(**{}, datasets=[dataset])
 
         with NamedTemporaryFile(mode="w+", delete=True) as fixtures_fd:
             # Get the fixtures from the local instance.
@@ -58,6 +61,7 @@ class FixturesTest:
         assert models.Discussion.objects.count() > 0
         assert models.CommunityResource.objects.count() > 0
         assert models.User.objects.count() > 0
+        assert models.Dataservice.objects.count() > 0
 
     def test_import_fixtures_from_default_file(self, cli):
         """Test importing fixtures from udata.commands.fixture.DEFAULT_FIXTURE_FILE."""
@@ -66,3 +70,5 @@ class FixturesTest:
         assert models.Dataset.objects.count() > 0
         assert models.Reuse.objects.count() > 0
         assert models.User.objects.count() > 0
+        if udata.commands.fixtures.DEFAULT_FIXTURE_FILE_TAG > "v1.0.0":
+            assert models.Dataservice.objects.count() > 0